Extended client api to accept setting for each executing query (this allows to set any clickhouse settings for a specific query)
