+++ client/src/main/scala/com.crobox.clickhouse/ClickhouseClient.scala
@@ -6,9 +6,9 @@ import akka.http.scaladsl.model._
 import akka.stream.scaladsl.{Framing, Source}
 import akka.util.ByteString
 import com.crobox.clickhouse.balancing.HostBalancer
-import com.crobox.clickhouse.internal.ClickHouseExecutor.QuerySettings
-import com.crobox.clickhouse.internal.ClickHouseExecutor.QuerySettings.{AllQueries, ReadQueries}
-import com.crobox.clickhouse.internal.{ClickHouseExecutor, ClickhouseQueryBuilder, ClickhouseResponseParser}
+import com.crobox.clickhouse.internal.QuerySettings._
+import com.crobox.clickhouse.internal.progress.QueryProgress.QueryProgress
+import com.crobox.clickhouse.internal.{ClickHouseExecutor, ClickhouseQueryBuilder, ClickhouseResponseParser, QuerySettings}
 import com.typesafe.config.Config
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -21,7 +21,7 @@ import scala.concurrent.{ExecutionContext, Future}
  * @since 31-03-17
  */
 class ClickhouseClient(override val config: Config, val database: String = "default")(
-    override implicit val system: ActorSystem = ActorSystem("clickhouseClient")
+    override implicit val system: ActorSystem = ActorSystem("clickhouseClient", config)
 ) extends ClickHouseExecutor
     with ClickhouseResponseParser
     with ClickhouseQueryBuilder {
@@ -47,17 +47,19 @@ class ClickhouseClient(override val config: Config, val database: String = "defa
    * @param sql a valid Clickhouse SQL string
    * @return Future with the result that clickhouse returns
    */
-  def query(sql: String): Future[String] =
-    executeRequest(sql, QuerySettings(ReadQueries))
+  def query(sql: String)(implicit settings: QuerySettings = QuerySettings(ReadQueries)): Future[String] =
+    executeRequest(sql, settings.copy(readOnly = ReadQueries))
 
   /**
    * Execute a read-only query on Clickhouse
-   *
+   * Experimental api, may change in the future.
    * @param sql a valid Clickhouse SQL string
    * @return stream with the query progress (started/rejected/finished/failed) which materializes with the query result
    */
-  def queryWithProgress(sql: String): Source[ClickHouseExecutor.QueryProgress, Future[String]] =
-    executeRequestWithProgress(sql, QuerySettings(ReadQueries))
+  def queryWithProgress(sql: String)(
+      implicit settings: QuerySettings = QuerySettings(ReadQueries)
+  ): Source[QueryProgress, Future[String]] =
+    executeRequestWithProgress(sql, settings.copy(readOnly = ReadQueries))
 
   /**
    * Execute a query that is modifying the state of the database. e.g. INSERT, SET, CREATE TABLE.
@@ -66,25 +68,26 @@ class ClickhouseClient(override val config: Config, val database: String = "defa
    * @param sql a valid Clickhouse SQL string
    * @return Future with the result that clickhouse returns
    */
-  def execute(sql: String): Future[String] =
+  def execute(sql: String)(implicit settings: QuerySettings = QuerySettings(AllQueries)): Future[String] =
     Future {
       require(
         !(sql.toUpperCase.startsWith("SELECT") || sql.toUpperCase.startsWith("SHOW")),
         ".execute() is not allowed for SELECT or SHOW statements, use .query() instead"
       )
     }.flatMap(
-      _ => executeRequest(sql, QuerySettings(AllQueries))
+      _ => executeRequest(sql, settings)
     )
 
-  def execute(sql: String, entity: String): Future[String] =
-    executeRequest(sql, QuerySettings(AllQueries), Option(entity))
+  def execute(sql: String,
+              entity: String)(implicit settings: QuerySettings): Future[String] =
+    executeRequest(sql, settings, Option(entity))
 
   /**
    * Creates a stream of the SQL query that will delimit the result from Clickhouse on new-line
    *
    * @param sql a valid Clickhouse SQL string
    */
-  def source(sql: String): Source[String, NotUsed] =
+  def source(sql: String)(implicit settings: QuerySettings = QuerySettings(ReadQueries)): Source[String, NotUsed] =
     sourceByteString(sql)
       .via(Framing.delimiter(ByteString("\n"), MaximumFrameLength))
       .map(_.utf8String)
@@ -94,10 +97,12 @@ class ClickhouseClient(override val config: Config, val database: String = "defa
    *
    * @param sql a valid Clickhouse SQL string
    */
-  def sourceByteString(sql: String): Source[ByteString, NotUsed] =
+  def sourceByteString(
+      sql: String
+  )(implicit settings: QuerySettings = QuerySettings(ReadQueries)): Source[ByteString, NotUsed] =
     Source
       .fromFuture(hostBalancer.nextHost.flatMap { host =>
-        singleRequest(toRequest(host, sql, None, QuerySettings(ReadQueries)))
+        singleRequest(toRequest(host, sql, None, settings, None)(config))
       })
       .flatMapConcat(response => response.entity.withoutSizeLimit().dataBytes)
 
@@ -108,9 +113,11 @@ class ClickhouseClient(override val config: Config, val database: String = "defa
    * @param source the Source with strings
    * @return Future with the result that clickhouse returns
    */
-  def sink(sql: String, source: Source[ByteString, Any]): Future[String] = {
+  def sink(sql: String, source: Source[ByteString, Any])(
+      implicit settings: QuerySettings = QuerySettings(AllQueries)
+  ): Future[String] = {
     val entity = HttpEntity.apply(ContentTypes.`text/plain(UTF-8)`, source)
-    executeRequest(sql, QuerySettings(AllQueries), Option(entity))
+    executeRequest(sql, settings, Option(entity))
   }
 
 }
