+++ client/src/main/scala/com.crobox.clickhouse/internal/ClickHouseExecutor.scala
@@ -3,21 +3,18 @@ package com.crobox.clickhouse.internal
 import akka.Done
 import akka.actor.ActorSystem
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.model.Uri.Query
 import akka.http.scaladsl.model._
 import akka.http.scaladsl.settings.{ClientConnectionSettings, ConnectionPoolSettings}
 import akka.stream._
-import akka.stream.scaladsl.{BroadcastHub, Keep, Sink, Source, SourceQueueWithComplete}
+import akka.stream.scaladsl.{Keep, Sink, Source, SourceQueueWithComplete}
 import com.crobox.clickhouse.balancing.HostBalancer
-import com.crobox.clickhouse.internal.ClickHouseExecutor.QuerySettings.ReadOnlySetting
-import com.crobox.clickhouse.internal.ClickHouseExecutor._
+import com.crobox.clickhouse.internal.progress.QueryProgress._
+import com.crobox.clickhouse.internal.progress.{QueryProgress, StreamingProgressClickhouseTransport}
 import com.typesafe.config.Config
 import com.typesafe.scalalogging.LazyLogging
 
 import scala.concurrent.{ExecutionContext, Future, Promise}
-import scala.util.parsing.json.JSON
-import scala.util.{Failure, Random, Success, Try}
-
+import scala.util.{Failure, Random, Success}
 private[clickhouse] trait ClickHouseExecutor extends LazyLogging {
   this: ClickhouseResponseParser with ClickhouseQueryBuilder =>
 
@@ -25,48 +22,10 @@ private[clickhouse] trait ClickHouseExecutor extends LazyLogging {
   protected implicit lazy val materializer: Materializer = ActorMaterializer()
   protected implicit val executionContext: ExecutionContext
   protected val hostBalancer: HostBalancer
-  protected def config: Config
+  protected val config: Config
 
   lazy val (progressQueue, progressSource) = {
-    val builtSource = Source
-      .queue[String](1000, OverflowStrategy.dropHead)
-      .map[Option[ClickhouseQueryProgress]](queryAndProgress => {
-        queryAndProgress.split("\n", 2).toList match {
-          case queryId :: ProgressHeadersAsEventsStage.AcceptedMark :: Nil =>
-            Some(ClickhouseQueryProgress(queryId, QueryAccepted))
-          case queryId :: progressJson :: Nil =>
-            Try {
-              val parsedJson = JSON.parseFull(progressJson).map(_.asInstanceOf[Map[String, String]])
-              if (parsedJson.isEmpty || parsedJson.get.size != 3) {
-                throw new IllegalArgumentException(s"Cannot extract progress from $parsedJson")
-              } else {
-                val jsonMap = parsedJson.get
-                ClickhouseQueryProgress(
-                  queryId,
-                  Progress(jsonMap("read_rows").toLong, jsonMap("read_bytes").toLong, jsonMap("total_rows").toLong)
-                )
-              }
-            } match {
-              case Success(value) => Some(value)
-              case Failure(exception) =>
-                logger.warn(s"Failed to parse json $progressJson", exception)
-                None
-            }
-          case other @ _ =>
-            logger.warn(s"Could not get progress from $other")
-            None
-
-        }
-      })
-      .collect {
-        case Some(progress) => progress
-      }
-      .withAttributes(ActorAttributes.supervisionStrategy({
-        case ex @ _ =>
-          logger.warn("Detected failure in the query progress stream, resuming operation.", ex)
-          Supervision.Resume
-      }))
-      .toMat(BroadcastHub.sink)(Keep.both)
+    val builtSource = QueryProgress.queryProgressStream
       .run()
     builtSource._2.runWith(Sink.ignore) //ensure we have one sink draining the progress
     builtSource
@@ -91,7 +50,7 @@ private[clickhouse] trait ClickHouseExecutor extends LazyLogging {
     })(Keep.both)
     .run
 
-  private val queryRetries: Int = config.getInt("crobox.clickhouse.client.retries")
+  private lazy val queryRetries: Int = config.getInt("crobox.clickhouse.client.retries")
 
   def executeRequest(query: String,
                      settings: QuerySettings,
@@ -155,12 +114,12 @@ private[clickhouse] trait ClickHouseExecutor extends LazyLogging {
       val request = toRequest(actualHost,
                               query,
                               Some(queryIdentifier),
-                              settings.withFallback(config),
-                              entity,
-                              progressQueue.isDefined)
+                              settings.copy(progressHeaders = Some(progressQueue.isDefined)),
+                              entity)(config)
       processClickhouseResponse(singleRequest(request), query, actualHost, progressQueue)
     })
   }
+
   private def executeWithRetries(retries: Int, progressQueue: Option[SourceQueueWithComplete[QueryProgress]])(
       request: () => Future[String]
   ): Future[String] =
@@ -178,67 +137,4 @@ private[clickhouse] trait ClickHouseExecutor extends LazyLogging {
     }
 }
 
-object ClickHouseExecutor {
-  val InternalQueryIdentifier = "X-Internal-Identifier"
-
-  sealed trait QueryProgress
-  case object QueryAccepted                                 extends QueryProgress
-  case object QueryFinished                                 extends QueryProgress
-  case object QueryRejected                                 extends QueryProgress
-  case class QueryFailed(cause: Throwable)                  extends QueryProgress
-  case class QueryRetry(cause: Throwable, retryNumber: Int) extends QueryProgress
-
-  case class ClickhouseQueryProgress(identifier: String, progress: QueryProgress)
-  case class Progress(rowsRead: Long, bytesRead: Long, totalRows: Long) extends QueryProgress
-
-  case class QuerySettings(readOnly: ReadOnlySetting,
-                           authentication: Option[(String, String)] = None,
-                           progressHeaders: Option[Boolean] = None,
-                           queryId: Option[String] = None,
-                           profile: Option[String] = None,
-                           httpCompression: Option[Boolean] = None) {
-
-    def asQueryParams: Query =
-      Query(
-        Seq("readonly"         -> readOnly.value.toString) ++
-        queryId.map("query_id" -> _) ++
-        authentication.map(
-          auth => "user" -> auth._1
-        ) ++
-        authentication.map(auth => "password" -> auth._2) ++
-        profile.map("profile" -> _) ++
-        progressHeaders.map(
-          progress => "send_progress_in_http_headers" -> (if (progress) "1" else "0")
-        ) ++
-        httpCompression.map(compression => "enable_http_compression" -> (if (compression) "1" else "0")): _*
-      )
-
-    def withFallback(config: Config): QuerySettings =
-      this.copy(
-        authentication = authentication.orElse(Try {
-          val authConfig = config.getConfig(path("authentication"))
-          (authConfig.getString("user"), authConfig.getString("password"))
-        }.toOption),
-        profile = profile.orElse(Try { config.getString(path("profile")) }.toOption),
-        httpCompression = httpCompression.orElse(Try { config.getBoolean(path("http-compression")) }.toOption)
-      )
-
-    private def path(setting: String) = s"crobox.clickhouse.client.settings.$setting"
-
-  }
-
-  object QuerySettings {
-    sealed trait ReadOnlySetting {
-      val value: Int
-    }
-    case object AllQueries extends ReadOnlySetting {
-      override val value: Int = 0
-    }
-    case object ReadQueries extends ReadOnlySetting {
-      override val value: Int = 1
-    }
-    case object ReadAndChangeQueries extends ReadOnlySetting {
-      override val value: Int = 2
-    }
-  }
-}
+object ClickHouseExecutor {}
