+++ client/src/main/scala/com.crobox.clickhouse/internal/ClickhouseQueryBuilder.scala
@@ -3,8 +3,8 @@ package com.crobox.clickhouse.internal
 import akka.http.scaladsl.model.Uri.Query
 import akka.http.scaladsl.model.headers.{HttpEncodingRange, RawHeader}
 import akka.http.scaladsl.model.{HttpMethods, HttpRequest, RequestEntity, Uri}
-import com.crobox.clickhouse.internal.ClickHouseExecutor.QuerySettings
-import com.crobox.clickhouse.internal.ClickHouseExecutor.QuerySettings.ReadQueries
+import com.crobox.clickhouse.internal.progress.ProgressHeadersAsEventsStage
+import com.typesafe.config.Config
 import com.typesafe.scalalogging.LazyLogging
 
 import scala.collection.immutable
@@ -20,10 +20,9 @@ private[clickhouse] trait ClickhouseQueryBuilder extends LazyLogging {
 
   protected def toRequest(uri: Uri,
                           query: String,
-                          queryIdentifier: Option[String] = None,
-                          settings: QuerySettings = QuerySettings(ReadQueries),
-                          entity: Option[RequestEntity] = None,
-                          progress: Boolean = false): HttpRequest =
+                          queryIdentifier: Option[String],
+                          settings: QuerySettings,
+                          entity: Option[RequestEntity])(config: Config): HttpRequest =
     entity match {
       case Some(e) =>
         logger.debug(s"Executing clickhouse query [$query] on host [${uri
@@ -31,18 +30,18 @@ private[clickhouse] trait ClickhouseQueryBuilder extends LazyLogging {
         HttpRequest(
           method = HttpMethods.POST,
           uri = uri.withQuery(
-            Query(Query("query" -> query) ++ settings.copy(progressHeaders = Some(progress)).asQueryParams: _*)
+            Query(Query("query" -> query) ++ settings.withFallback(config).asQueryParams: _*)
           ),
           entity = e,
-          headers = Headers ++ queryIdentifier.map(RawHeader(ClickHouseExecutor.InternalQueryIdentifier, _))
+          headers = Headers ++ queryIdentifier.map(RawHeader(ProgressHeadersAsEventsStage.InternalQueryIdentifier, _))
         )
       case None =>
         logger.debug(s"Executing clickhouse query [$query] on host [${uri.toString()}]")
         HttpRequest(
           method = HttpMethods.POST,
-          uri = uri.withQuery(settings.copy(progressHeaders = Some(progress)).asQueryParams),
+          uri = uri.withQuery(settings.withFallback(config).asQueryParams),
           entity = query,
-          headers = Headers ++ queryIdentifier.map(RawHeader(ClickHouseExecutor.InternalQueryIdentifier, _))
+          headers = Headers ++ queryIdentifier.map(RawHeader(ProgressHeadersAsEventsStage.InternalQueryIdentifier, _))
         )
     }
