+++ client/src/test/scala/com/crobox/clickhouse/internal/ClickhouseExecutorTest.scala
@@ -5,34 +5,36 @@ import akka.stream.scaladsl.{Sink, SourceQueue}
 import akka.stream.{Materializer, StreamTcpException}
 import com.crobox.clickhouse.ClickhouseClientAsyncSpec
 import com.crobox.clickhouse.balancing.HostBalancer
-import com.crobox.clickhouse.internal.ClickHouseExecutor.QuerySettings.ReadQueries
-import com.crobox.clickhouse.internal.ClickHouseExecutor.{QueryProgress, QueryRetry, QuerySettings}
+import com.crobox.clickhouse.internal.QuerySettings.ReadQueries
+import com.crobox.clickhouse.internal.progress.QueryProgress.{QueryProgress, QueryRetry}
 import com.typesafe.config.Config
 
 import scala.concurrent.{ExecutionContext, Future}
 
 class ClickhouseExecutorTest extends ClickhouseClientAsyncSpec {
-  private val self                     = this
+  private lazy val self                = this
   private var response: Future[String] = Future.successful("")
-  private val executor = new ClickHouseExecutor with ClickhouseResponseParser with ClickhouseQueryBuilder {
-    override protected implicit val system: ActorSystem                = self.system
-    override protected implicit val executionContext: ExecutionContext = system.dispatcher
-    override protected val hostBalancer: HostBalancer                  = HostBalancer(config)
-    override protected def config: Config                              = self.config
-    override protected def processClickhouseResponse(
-        responseFuture: Future[HttpResponse],
-        query: String,
-        host: Uri,
-        progressQueue: Option[
-          SourceQueue[
-            ClickHouseExecutor.QueryProgress
+  private lazy val executor = {
+    new ClickHouseExecutor with ClickhouseResponseParser with ClickhouseQueryBuilder {
+      override protected implicit val system: ActorSystem                = self.system
+      override protected implicit val executionContext: ExecutionContext = system.dispatcher
+      override protected val config: Config                              = self.config
+      override protected val hostBalancer: HostBalancer                  = HostBalancer(config)
+      override protected def processClickhouseResponse(
+          responseFuture: Future[HttpResponse],
+          query: String,
+          host: Uri,
+          progressQueue: Option[
+            SourceQueue[
+              QueryProgress
+            ]
           ]
-        ]
-    )(
-        implicit materializer: Materializer,
-        executionContext: ExecutionContext
-    ): Future[String] =
-      response
+      )(
+          implicit materializer: Materializer,
+          executionContext: ExecutionContext
+      ): Future[String] =
+        response
+    }
   }
 
   it should "retry requests" in {
